Move StringBufferType to BufferType and use labels
Refactor buffer type handling by moving the StringBufferType enum out of APIData into src/model/BufferType, and add a StringBufferTypeLabel mapping for human-readable labels. Update APIData interfaces to reference the new StringBufferType type, and adjust imports/usages across components and functions (DeviceOperationsFullRender, MemoryLegendElement, processMemoryAllocations, OperationDetails) to use the centralized enum and label mapping. This centralizes buffer type definitions and simplifies rendering of buffer type labels in the UI.
