+++ src/components/operation-details/DeviceOperationsFullRender.tsx
@@ -21,15 +21,7 @@ import { useAtomValue } from 'jotai';
 import classNames from 'classnames';
 import hljs from 'highlight.js/lib/core';
 import 'highlight.js/styles/a11y-dark.css';
-import {
-    BufferNode,
-    DeviceOperationNode,
-    Node,
-    NodeType,
-    StringBufferType,
-    Tensor,
-    TensorNode,
-} from '../../model/APIData';
+import { BufferNode, DeviceOperationNode, Node, NodeType, Tensor, TensorNode } from '../../model/APIData';
 import 'styles/components/DeviceOperationFullRender.scss';
 import 'styles/components/DeviceOperationArgumentsComponent.scss';
 import { MemoryLegendElement } from './MemoryLegendElement';
@@ -42,7 +34,7 @@ import { L1_DEFAULT_MEMORY_SIZE, L1_NUM_CORES } from '../../definitions/L1Memory
 import { getBufferColor, getTensorColor } from '../../functions/colorGenerator';
 import MemoryTag from '../MemoryTag';
 import { toReadableLayout, toReadableShape } from '../../functions/formatting';
-import { BufferTypeLabel } from '../../model/BufferType';
+import { BufferTypeLabel, StringBufferType } from '../../model/BufferType';
 
 type BufferDetails = {
     bufferOrTensorNode?: BufferNode | TensorNode;
@@ -69,7 +61,7 @@ const renderBufferDetails = ({ bufferOrTensorNode, tensorId, optionalOutput, det
         const { params } = bufferOrTensorNode;
         address = parseInt(params.address, 10);
         layout = toReadableLayout(params.layout) || '';
-        type = BufferTypeLabel[params.buffer_type] as StringBufferType;
+        type = BufferTypeLabel[params.buffer_type] as StringBufferType; // converting int enum to string representation
     }
 
     if (address !== undefined || tensorId !== undefined) {
