+++ src/model/APIData.ts
@@ -5,7 +5,7 @@
 import { RemoteConnection, RemoteFolder } from '../definitions/RemoteConnection';
 import { ReportLocation } from '../definitions/Reports';
 import { BufferMemoryLayout, MemoryConfig } from '../functions/parseMemoryConfig';
-import { BufferType } from './BufferType';
+import { BufferType, StringBufferType } from './BufferType';
 
 interface OperationError {
     operation_id: number;
@@ -239,14 +239,6 @@ export enum DeviceOperationLayoutTypes {
     TILE = 'TILE',
 }
 
-export enum StringBufferType {
-    DRAM = 'DRAM',
-    L1 = 'L1',
-    SYSTEM_MEMORY = 'SYSTEM MEMORY',
-    L1_SMALL = 'L1 SMALL',
-    TRACE = 'TRACE',
-}
-
 export interface DeviceOperationParams {
     name: string;
     device_id?: number | string;
@@ -266,7 +258,7 @@ interface BufferAllocateParams {
     num_cores: string; // '64';
     page_size: string; // '448';
     size: string; // '7340032';
-    type: keyof typeof StringBufferType; // 'L1';
+    type: StringBufferType; // 'L1';
     derived_device_id?: number[];
 }
 interface CircularBufferAllocateParams {
@@ -289,7 +281,7 @@ export interface DeviceTensorParams {
     shape: string; // 'Shape([16, 3, 224, 224])';
     tensor_id: number; // '0';
     size: string; // '602112';
-    type: keyof typeof StringBufferType;
+    type: StringBufferType;
 }
 
 export interface BaseNode<T extends NodeType, P> {
@@ -357,7 +349,7 @@ export interface CircularBuffer extends Chunk {
 
 export interface TensorBuffer extends Chunk {
     layout: DeviceOperationLayoutTypes;
-    type: keyof typeof StringBufferType;
+    type: StringBufferType;
 }
 
 export interface BufferPage {
